+++ packages/hono/src/response/json/index.ts
@@ -1,9 +1,9 @@
-import type { CreateJsonResponseStructOptions } from "@jderjs/core/internal";
+import type { CreateJsonResponseStructOptions } from "@jderjs/core/response/json/struct";
 import type { Context } from "hono";
 import type { StatusCode } from "hono/utils/http-status";
 import type { Format, Omit } from "ts-vista";
 
-import { createJsonResponseStruct } from "@jderjs/core/internal";
+import { createJsonResponseStruct } from "@jderjs/core/response/json/struct";
 
 /** Check if the argument is a Hono context. */
 const isContext = (arg: any): arg is Context => arg?.req !== undefined;